fix: sort function error * refactor: new getERC721TokenDetailed file * fix: i18n
