+++ packages/maskbook/src/plugins/Election2020/UI/ElectionTokenAlbum.tsx
@@ -1,11 +1,10 @@
 import { createStyles, makeStyles, Typography } from '@material-ui/core'
 import { useConstant } from '../../../web3/hooks/useConstant'
-import { getERC721TokenDetailed } from '../../../web3/hooks/useERC721TokenDetailed'
 import { ELECTION_2020_CONSTANTS } from '../constants'
 import { useAllElectionTokensOfOwner } from '../hooks/useAllElectionTokensOfOwner'
 import { ElectionCard } from './ElectionCard'
 import { useI18N } from '../../../utils/i18n-next-ui'
-import { Suspense } from 'react'
+import { getERC721TokenDetailed } from '../../../web3/suspends/getERC721TokenDetailed'
 
 const useStyles = makeStyles((theme) =>
     createStyles({
